use size_t instead
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@3613 a1c6a512-1295-4272-9138-f99709370657

+++ uisimulator/x11/sprintf.h
@@ -1,7 +1,7 @@
 #include <stdarg.h>
 #include <stdio.h>
 
-int rockbox_snprintf (char *buf, int size, const char *fmt, ...);
+int rockbox_snprintf (char *buf, size_t size, const char *fmt, ...);
 int rockbox_vsnprintf (char *buf, int size, const char *fmt, va_list ap);
 int rockbox_fprintf (int fd, const char *fmt, ...);
